now with 100% more fuzzy

+++ main.py
@@ -7,6 +7,8 @@ Inputs:
 - PAT of appropriate scope (assumes the workflow token if not specified)
 - Report scope ("enterprise", "organization", "repository")
 - Enterprise slug OR organization name OR repository name
+- Dry run (False if not set, True if set to literally anything)
+- Substring (string to match against runner name, such as to only delete runners with "test" in the name)
 
 Outputs:
 - Nothing really, it removes your offline runners
@@ -41,6 +43,11 @@ if os.environ.get("DRY_RUN") is None:
     dry_run = False
 else:
     dry_run = True
+
+if os.environ.get("FUZZY_NAME") is None:
+    substring = ""
+else:
+    substring = os.environ.get("FUZZY_NAME")
 
 # Define functions
 def set_url(api_endpoint, runner_scope, scope_name):
@@ -83,17 +90,17 @@ def get_runners(base_url, headers):
     return runner_list
 
 
-def delete_runners(base_url, headers, runner_list, dry_run):
+def delete_runners(base_url, headers, runner_list, dry_run, substring):
     """
     Delete the offline runners
     """
     for i in runner_list:
-        if i["status"] == "offline" and dry_run == False:
+        if i["status"] == "offline" and substring in i["name"] and dry_run == False:
             url = base_url + "/{}".format(i["id"])
             response = requests.delete(url, headers=headers)
             if response.status_code == 204:
                 print("Deleted runner {}".format(i["name"]))
-        elif i["status"] == "offline" and dry_run == True:
+        elif i["status"] == "offline" and substring in i["name"] and dry_run == True:
             print("Runner {} is offline and would be deleted".format(i["name"]))
 
 
@@ -106,4 +113,4 @@ if __name__ == "__main__":
     }
     base_url = set_url(api_endpoint, runner_scope, scope_name)
     runner_list = get_runners(base_url, headers)
-    delete_runners(base_url, headers, runner_list, dry_run)
+    delete_runners(base_url, headers, runner_list, dry_run, substring)
